+++ lib/gg_git_test_helpers.dart
@@ -1,0 +1,7 @@
+// @license
+// Copyright (c) 2019 - 2024 Dr. Gabriel Gatzsche. All Rights Reserved.
+//
+// Use of this source code is governed by terms that can be
+// found in the LICENSE file in the root of this package.
+
+export 'src/test_helpers/test_helpers.dart';