+++ lib/src/test_helpers/test_helpers.dart
@@ -4,16 +4,12 @@
 // Use of this source code is governed by terms that can be
 // found in the LICENSE file in the root of this package.
 
-// .............................................................................
 import 'dart:io';
 
-// .............................................................................
-Future<void> writeFile(File file, String content) async {
-  file.writeAsStringSync(content, flush: true);
-  // await Future<void>.delayed(const Duration(microseconds: 5));
-}
+// coverage:ignore-file
 
 // .............................................................................
+/// Initializes a test directory
 Directory initTestDir() {
   final tmpBase =
       Directory('/tmp').existsSync() ? Directory('/tmp') : Directory.systemTemp;
@@ -30,6 +26,7 @@ Directory initTestDir() {
 }
 
 // .............................................................................
+/// Init git repository in test directory
 Future<void> initGit(Directory testDir) async {
   final result =
       await Process.run('git', ['init'], workingDirectory: testDir.path);
@@ -39,6 +36,7 @@ Future<void> initGit(Directory testDir) async {
 }
 
 // .............................................................................
+/// Init remote git repository in directory
 Directory initRemoteGit(Directory testDir) {
   final remoteDir = Directory('${testDir.path}/remote');
   remoteDir.createSync(recursive: true);
@@ -55,6 +53,7 @@ Directory initRemoteGit(Directory testDir) {
 }
 
 // .............................................................................
+/// Init local git repository in directory
 Directory initLocalGit(Directory testDir) {
   final localDir = Directory('${testDir.path}/local');
   localDir.createSync(recursive: true);
@@ -75,6 +74,7 @@ Directory initLocalGit(Directory testDir) {
 // # Tag helpers
 // #############
 
+/// Add tag to test directory
 Future<void> addTag(Directory testDir, String tag) async {
   final result = await Process.run(
     'git',
@@ -86,6 +86,7 @@ Future<void> addTag(Directory testDir, String tag) async {
   }
 }
 
+/// Add tags to test directory
 Future<void> addTags(Directory testDir, List<String> tags) async {
   for (final tag in tags) {
     await addTag(testDir, tag);
@@ -97,10 +98,12 @@ Future<void> addTags(Directory testDir, List<String> tags) async {
 // ##############
 
 // .............................................................................
+/// Init a file with a name in the test directory
 Future<void> initFile(Directory testDir, String name, String content) =>
-    writeFile(File('${testDir.path}/$name'), content);
+    File('${testDir.path}/$name').writeAsString(content);
 
 // .............................................................................
+/// Commit the file with a name in the test directory
 void commitFile(Directory testDir, String name) {
   final result = Process.runSync(
     'git',
@@ -123,29 +126,33 @@ void commitFile(Directory testDir, String name) {
 // ## sample.txt
 
 // .............................................................................
+/// Add and commit sample file
 void addAndCommitSampleFile(Directory testDir) {
   initFile(testDir, 'sample.txt', 'sample');
   commitFile(testDir, 'sample.txt');
 }
 
 // .............................................................................
+/// Update and commit sample file
 Future<void> updateAndCommitSampleFile(Directory testDir) async {
   final file = File('${testDir.path}/sample.txt');
   final content = file.existsSync() ? file.readAsStringSync() : '';
   final newContent = '${content}updated';
-  await writeFile(File('${testDir.path}/sample.txt'), newContent);
+  await File('${testDir.path}/sample.txt').writeAsString(newContent);
   commitFile(testDir, 'sample.txt');
 }
 
 // ## uncommitted.txt
 
 // .............................................................................
+/// Init uncommitted file
 void initUncommitedFile(Directory testDir) =>
     initFile(testDir, 'uncommitted.txt', 'uncommitted');
 
 // ## pubspect.yaml
 
 // .............................................................................
+/// Create a pubspec.yaml file with a version
 Future<void> setPubspec(Directory testDir, {required String? version}) async {
   final file = File('${testDir.path}/pubspec.yaml');
 
@@ -159,15 +166,17 @@ Future<void> setPubspec(Directory testDir, {required String? version}) async {
     content = content.replaceAll(RegExp(r'version: .*'), 'version: $version');
   }
 
-  await writeFile(file, content);
+  await file.writeAsString(content);
 }
 
 // .............................................................................
+/// Commit the pubspec file
 void commitPubspec(Directory testDir) => commitFile(testDir, 'pubspec.yaml');
 
 // ## CHANGELOG.md
 
 // .............................................................................
+/// Create a CHANGELOG.md file with a version
 Future<void> setChangeLog(
   Directory testDir, {
   required String? version,
@@ -180,9 +189,14 @@ Future<void> setChangeLog(
   await initFile(testDir, 'CHANGELOG.md', content);
 }
 
+// .............................................................................
+/// Commit the changelog file
+void commitChangeLog(Directory testDir) => commitFile(testDir, 'CHANGELOG.md');
+
 // ## Version files
 
 // .............................................................................
+/// Write version into pubspec.yaml, Changelog.md and add a tag
 Future<void> setupVersions(
   Directory testDir, {
   required String? pubspec,
@@ -198,6 +212,3 @@ Future<void> setupVersions(
     await addTag(testDir, gitHead);
   }
 }
-
-// .............................................................................
-void commitChangeLog(Directory testDir) => commitFile(testDir, 'CHANGELOG.md');