+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1
 
 - Update dependencies
+- Make test_helpers public
 
 ## 1.0.0